+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ---
 - Added example notebooks
 - Added ToolPreview redirect
+- Fixed Python 3 support
 - Replaced pgrep with psutil to support non-POSIX systems
 
 0.1