+++ CHANGES.rst
@@ -1,5 +1,6 @@
 0.4
 ---
+- Improve unicode support
 - Save tool template if first cell is markdown
 - Save result template if last cell is markdown